+++ CHANGELOG.md
@@ -32,6 +32,28 @@ This document details all notable changes to the WSU HRS Child Theme. Uses [Sema
 - Get Babel processing working, open #35.
 - Determine if we really need the `center-middle` utility class. If yes, look into using a flex or grid with `Xvh` and `Xvh` column and row size to center instead.
 
+## 0.12.0 (2018-06-08)
+
+### Changed
+
+- Clean up `functions.php` to remove unneeded methods and refactor the rest.
+- Prefer to enqueue the login page CSS rather than inline it.
+- Rename HRS nav menus for easier identification.
+- Move WP menu(s) registration into the setup class.
+- Fix malformed call to `add_theme_support` for HTML5 search form and move to the setup class.
+
+### Added
+
+- Filter to adjust the punctuation in the HRS page title element.
+- Crimson svg WSU cougar logo for CSS background use.
+- New nav menu to display the site footer nav menu so that the "Offsite" menu can display in the Spine as intended, or not at all if unwanted.
+- Action and methods to get and set the Spine schema version for the HRS child theme -- set to 2.x to target the latest Spine configuration options.
+- A theme setup class to handle theme setup tasks such as registering theme support in a new `includes/` directory.
+
+### Removed
+
+- Restored the "lost password" link to the login page by removing the filter that erased it.
+
 ## 0.11.1 (2018-06-05)
 
 ### Fixed